% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iisgcolors.R
\name{iisgcolors}
\alias{iisgcolors}
\title{Vectors of colors for figures
Creates different vectors of related colors that may be useful for various IISG-related figures}
\usage{
iisgcolors(set = c("main", "blues", "purples", "teals", "oranges"))
}
\arguments{
\item{set}{Character string indicating a set of colors}
}
\value{
Vector of character strings
}
\description{
Vectors of colors for figures
Creates different vectors of related colors that may be useful for various IISG-related figures
}
\examples{
plot(0, 0, type="n", xlab="", ylab="", xlim=c(0, 9), ylim=c(7.5, 0), yaxs="i",
     xaxt="n", yaxt="n", mar=c(0.6, 5.1, 0.6, 0.6), xaxs="i")
axis(side=2, at=1:5, c("main", "blues", "purples", "teals", "oranges"), las=1)

gen <- iisgcolors("main")
points(seq(along=gen), rep(1,length(gen)), pch=21, bg=gen, cex=4)
text(seq(along=gen), rep(c(0.55, 0.7), length(gen))[seq(along=gen)], names(gen))

points(1, 2, pch=21, bg=iisgcolors("blues"), cex=4)
points(1, 3, pch=21, bg=iisgcolors("purples"), cex=4)

teals <- iisgcolors("teals")
points(seq(along=teals), rep(4,length(teals)), pch=21, bg=teals, cex=4)
text(seq(along=teals), rep(3+c(0.55, 0.7), length(teals))[seq(along=teals)], names(teals))

oranges <- iisgcolors("oranges")
points(seq(along=oranges), rep(5,length(oranges)), pch=21, bg=oranges, cex=4)
text(seq(along=oranges), rep(4.7, length(oranges)), names(oranges))

}
\author{
J Stuart Carlton, \email{stuart@jscarlton.net}
}
\keyword{utilities}
